zfs rig: zfs nas is now fully setup and deployed
I've fully deployed and setup the zfs nas on actual hardware. This has
allowed me to generate the flake.lock and hardware-configuration.nix
files and also allowed me to test and fix a few issues with the configs.

I've also updated the zfs.nix with the mountpoints and the dataset
structures of my NAS.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>