ssh & zfs: changed mounts & disabled boot decryption
I've disabled the decryption of datasets on boot and also removed any
encrypted drives from the auto-mount list so that the server doesn't
hang and wait for user input whenever it boots (which cannot be done over
SSH).

It also leaves the datasets encrypted unless explicitly decrypted by myself
which is much more secure.

I've also changed the SSH port to not conflict with anything else on the
network.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>