networking: added specialized firewall configs
I've updated the firewall to be more explicit rules and also to allow
any future minecraft servers to work aswell (port 25565).

I've also turned off fail2ban as it's incredibly annoying as the person
that gets banned most is me when I put in my password wrong. And plus
I'll turn it on if I need.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>