simple configurations for uncle's nas
I've added the basic configurations for uncle's nas.

I've decided to only give him one user besides the root user, that is
the 'server' user and I've also generated the hardware configurations
for the laptop we'll be using for the nas.

So far the drives and drive enclosure haven't arrived yet, so I haven't
made any real zfs configurations or setup scripts.

For ssh it uses port 22 now, as it makes things easier.

I've also changed the network settings to use network manager (for wifi)
instead as this server has no ethernet.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>