tailscale: added tailscale to bypass port fwding
I've added tailscale as a solution to get around the fact that port
forwarding is not possible on the network that I am setting this up for.
I've followed the standard tailscale instructions to get everything
working and it should work fine.
I've also used a imperatively created text file to store the tailscale
oneshot authentication key as I don't want to leak it in this git
commit and I also don't want to use some complicated solution like
nix-sops.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>